add token table UI and token row components
